Refactor: Tinkered with reward system some more
[Changes of note]
- Agent needs changed from type float to int

- Rework world bounds observations: Now relative to agent distance from border.

- Agent can now move diagonally across the grid.

- Agent need will refill if standing on object or any 8 surrounding gird positions.

- Add button in tilemap inspector to reset tilemap bounds.

+++ Assets/Scripts/StatusBar.cs
@@ -8,15 +8,15 @@ public class StatusBar : MonoBehaviour
     public Slider statusBar;
     public Image statusBarFill;
 
-    private int previousValue;
+    private float previousValue;
 
-    public void SetMaxValue(int statusValue)
+    public void SetMaxValue(float statusValue)
     {
         statusBar.maxValue = statusValue;
         statusBar.value = statusValue;
     }
 
-    public void SetValue(int statusValue)
+    public void SetValue(float statusValue)
     {
         // Change the colour of the fill if the value is increasing
         if (statusValue > previousValue)
